refactor: move some nodes to Inventor folder
