+++ src/Gui/CommandView.cpp
@@ -65,6 +65,7 @@
 #include "Document.h"
 #include "FileDialog.h"
 #include "ImageView.h"
+#include "Inventor/SoAxisCrossKit.h"
 #include "Macro.h"
 #include "MainWindow.h"
 #include "NavigationStyle.h"
@@ -73,7 +74,6 @@
 #include "SceneInspector.h"
 #include "Selection.h"
 #include "SelectionObject.h"
-#include "SoAxisCrossKit.h"
 #include "SoFCOffscreenRenderer.h"
 #include "TextureMapping.h"
 #include "Tools.h"
