+++ src/Gui/Inventor/SoDrawingGrid.cpp
@@ -23,7 +23,6 @@
 #include "PreCompiled.h"
 
 #ifndef _PreComp_
-# include <QObject>
 # include <Inventor/actions/SoGLRenderAction.h>
 # include <Inventor/elements/SoCacheElement.h>
 # include <Inventor/elements/SoLazyElement.h>
@@ -31,12 +30,11 @@
 # include <Inventor/elements/SoProjectionMatrixElement.h>
 # include <Inventor/elements/SoViewingMatrixElement.h>
 # include <Inventor/elements/SoViewportRegionElement.h>
-#endif
-
 #ifdef FC_OS_MACOSX
 #include <OpenGL/gl.h>
 #else
 #include <GL/gl.h>
+#endif
 #endif
 
 #include "SoDrawingGrid.h"
@@ -179,14 +177,14 @@ SoDrawingGrid::GLRenderOffPath(SoGLRenderAction *)
 void
 SoDrawingGrid::generatePrimitives(SoAction* action)
 {
-    Q_UNUSED(action); 
+    (void)action; 
 }
 
 void
 SoDrawingGrid::computeBBox(SoAction *action, SbBox3f &box, SbVec3f &center)
 {
-    Q_UNUSED(action); 
-    Q_UNUSED(box); 
-    Q_UNUSED(center); 
+    (void)action; 
+    (void)box; 
+    (void)center; 
     //SoState*  state = action->getState();
 }
