+++ src/Gui/View3DInventorViewer.cpp
@@ -94,14 +94,14 @@
 #include "Application.h"
 #include "Document.h"
 #include "GLPainter.h"
+#include "Inventor/SoAxisCrossKit.h"
+#include "Inventor/SoFCBoundingBox.h"
 #include "MainWindow.h"
 #include "Multisample.h"
 #include "NaviCube.h"
 #include "NavigationStyle.h"
 #include "Selection.h"
-#include "SoAxisCrossKit.h"
 #include "SoFCBackgroundGradient.h"
-#include "SoFCBoundingBox.h"
 #include "SoFCDB.h"
 #include "SoFCInteractiveElement.h"
 #include "SoFCOffscreenRenderer.h"
