+++ src/Gui/View3DPy.cpp
@@ -55,6 +55,7 @@
 
 #include "Camera.h"
 #include "Document.h"
+#include "Inventor/SoMouseWheelEvent.h"
 #include "NavigationStyle.h"
 #include "PythonWrapper.h"
 #include "SoFCDB.h"
@@ -62,7 +63,6 @@
 #include "SoFCSelectionAction.h"
 #include "SoFCVectorizeSVGAction.h"
 #include "SoFCVectorizeU3DAction.h"
-#include "SoMouseWheelEvent.h"
 #include "View3DInventor.h"
 #include "View3DInventorViewer.h"
 #include "ViewProviderDocumentObject.h"
