+++ src/Mod/Part/Gui/ViewProviderGridExtension.cpp
@@ -50,7 +50,7 @@
 #include <Gui/ViewProvider.h>
 #include <Gui/ViewProviderDocumentObject.h>
 
-#include <Gui/SoFCBoundingBox.h>
+#include <Gui/Inventor/SoFCBoundingBox.h>
 
 #include "ViewProviderGridExtension.h"
