+++ src/Mod/Sketcher/Gui/EditModeCoinManager.cpp
@@ -44,7 +44,7 @@
 
 #include <Base/Exception.h>
 #include <Gui/Inventor/MarkerBitmaps.h>
-#include <Gui/SoFCBoundingBox.h>
+#include <Gui/Inventor/SoFCBoundingBox.h>
 #include <Mod/Sketcher/App/Constraint.h>
 #include <Mod/Sketcher/App/GeoList.h>
